+++ packages/atomic/src/components/atomic-search-interface/atomic-search-interface.tsx
@@ -16,6 +16,9 @@ import {
   AnalyticsActions,
   ConfigurationActions,
   LogLevel,
+  buildSearchParameterManager,
+  buildSearchParameterSerializer,
+  Unsubscribe,
 } from '@coveo/headless';
 import {RenderError} from '../../utils/render-utils';
 import {InitializeEvent} from '../../utils/initialization-utils';
@@ -33,6 +36,7 @@ export class AtomicSearchInterface {
   @RenderError() error?: Error;
 
   private engine?: Engine;
+  private unsubscribe: Unsubscribe = () => {};
   private hangingComponentsInitialization: InitializeEvent[] = [];
   private initialized = false;
 
@@ -40,6 +44,10 @@ export class AtomicSearchInterface {
     if (this.sample) {
       this.initialize(HeadlessEngine.getSampleConfiguration());
     }
+  }
+
+  disconnectedCallback() {
+    this.unsubscribe();
   }
 
   @Method() async initialize(
@@ -84,13 +92,27 @@ export class AtomicSearchInterface {
     this.hangingComponentsInitialization = [];
 
     // Waits until the fields are registered asynchronously before triggering a search
-    setTimeout(
-      () =>
-        this.engine!.dispatch(
-          SearchActions.executeSearch(AnalyticsActions.logInterfaceLoad())
-        ),
-      0
-    );
+    setTimeout(() => {
+      this.initSearchParameterManager();
+
+      this.engine!.dispatch(
+        SearchActions.executeSearch(AnalyticsActions.logInterfaceLoad())
+      );
+    }, 0);
+  }
+
+  private initSearchParameterManager() {
+    const stateWithoutHashSign = window.location.hash.slice(1);
+    const {serialize, deserialize} = buildSearchParameterSerializer();
+    const params = deserialize(stateWithoutHashSign);
+
+    const manager = buildSearchParameterManager(this.engine!, {
+      initialState: {parameters: params},
+    });
+
+    this.unsubscribe = manager.subscribe(() => {
+      window.location.hash = serialize(manager.state.parameters);
+    });
   }
 
   @Watch('searchHub')